Added the rearranged refactor
