+++ src/include/GameManager.h
