+++ src/include/GameState.h
@@ -31,9 +31,23 @@ private:
       gameFlags(),
       pieceTextures(),
       moveCount(1),
-     positionHistory(),
-    selectedPieceIndex(-1) 
-    {} 
+      positionHistory(),
+      selectedPieceIndex(-1) 
+    {
+        // Initialize bitboards explicitly
+        bitboards.WhitePawns = 0ULL;
+        bitboards.BlackPawns = 0ULL;
+        bitboards.WhiteKnights = 0ULL;
+        bitboards.BlackKnights = 0ULL;
+        bitboards.WhiteBishops = 0ULL;
+        bitboards.BlackBishops = 0ULL;
+        bitboards.WhiteRooks = 0ULL;
+        bitboards.BlackRooks = 0ULL;
+        bitboards.WhiteQueens = 0ULL;
+        bitboards.BlackQueens = 0ULL;
+        bitboards.WhiteKing = 0ULL;
+        bitboards.BlackKing = 0ULL;
+    }
 
     GameState(const GameState&) = delete;
     GameState& operator=(const GameState&) = delete;
