+++ src/prog_chess_engine.cpp
@@ -1,4 +1,4 @@
-#include "GameManager.h"
+#include "include/GameManager.h"
 
 int main(void) {
     GameManager gameManager;
